wip usize accumulator bits

+++ src/lib.rs
@@ -12,6 +12,7 @@ use smart_leds_trait::{SmartLedsWriteAsync, RGB8};
 
 
 /// Write the colors to SPI bits in a buffer for transmission, returning the number of words.
+#[inline(never)]
 fn write_buffer<Word, Lut, T, I>(lut: &Lut, iterator: T, buffer: &mut [Word]) -> usize 
 where
     Lut: SmartLedsSpiData,
@@ -39,23 +40,23 @@ where
     };
 
     let mut accumulator: u32 = 0;  // current SPI data being built, MSbit aligned
-    let mut accumulator_bits: u8 = 0;  // number of bits written into accumulator
+    let mut accumulator_bits: usize = 0;  // number of bits written into accumulator
 
     for item in iterator {
         let item = item.into();
         let mut color_data = (item.g as u32) << 16 | (item.r as u32) << 8 | (item.b as u32);
 
         for _ in 0..(24 / Lut::INDEX_BITS) {
-            let lut_index = (color_data >> (24 - Lut::INDEX_BITS)) as u8;
+            let lut_index = (color_data >> (24 - Lut::INDEX_BITS)) as usize;
             let (spi_data, spi_bits) = lut.get(lut_index);
             color_data = color_data << Lut::INDEX_BITS;  // shift to get the next bits to the MSbits
 
             accumulator = accumulator | (spi_data >> accumulator_bits);
             accumulator_bits += spi_bits;
-            while accumulator_bits >= Word::BITS {
+            while accumulator_bits >= Word::BITS as usize {
                 push_word(Word::truncate_from_u32(accumulator >> accumulator_to_word_shift));
                 accumulator = accumulator << Word::BITS;
-                accumulator_bits -= Word::BITS;
+                accumulator_bits -= Word::BITS as usize;
             }
         }
     }
